Update move files
